code/requirements.txt
llama-cpp-python # llama.cpp python package
requests # used to download models